Simplify Stitching API
